Support multiple inputs in a single validator-wrapper (#1)
Having more than one input element is foreseeable and a norm. Although it can also significantly increase the complexity for the user--- both model and validator now it needs to be aware of the structure of the inner component. however, overall I believe this feature is required and can boost the productivity for form item does has more than one input.

after this change, the wrapper yields on the entire model object, the key-value structure represents the value and the associated element name, e.g.
```
{ 'email': 'abc.cde.com' }
```

similarly, the error becomes an object (from a string) that allows the consumer to manage separately.

* enable structured error object, this is essential for supporting multiple inputs within a single v-wrapper
* create @validate param for a single custom validator (simplify the usability)
* created `intersection` as a shared util
* create a getter for `validators`
* refine jsdoc based on the latest change
- fix eslint errors (exclude v-container)