Refactor for modifier (#4)
the current logic was relying on `ember/render-modifier`, which was an early
version of modifier and meant to polyfill the function gap for developer just
switch to glimmer from ember component. and the `render-modifier` has been
marked as `deprecated` from the recent version of Ember. the recommendation is
to migrate to ember-modifier.